# docs/api/openapi-bundled.yaml
openapi: 3.1.0
info:
  title: AuthFramework REST API
  description: |
    Comprehensive REST API for the AuthFramework authentication and authorization library.

    This API provides complete access to AuthFramework's authentication, authorization,
    user management, multi-factor authentication, OAuth 2.0, OIDC, RBAC, and administrative 
    features through standardized HTTP endpoints.

    ## Version Strategy

    - **Versioned endpoints** (`/api/v1/*`): All functional API endpoints are versioned
      for stability and backward compatibility. Version changes follow semantic versioning.

    - **Unversioned endpoints**: Infrastructure endpoints (`/health`, `/metrics`, `/readiness`, 
      `/liveness`) and OIDC discovery (`/.well-known/*`) remain unversioned for operational 
      stability and spec compliance.

    ## Authentication

    Most endpoints require authentication via Bearer tokens in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```

    Obtain tokens via:
    - `/api/v1/auth/login` - Username/password authentication
    - `/api/v1/oauth/token` - OAuth 2.0 token endpoint
    - `/api/v1/auth/refresh` - Refresh token flow

    ## Rate Limiting

    API endpoints are rate-limited to prevent abuse:
    - Authentication endpoints: 5 requests per minute per IP
    - Standard endpoints: 100 requests per minute per user
    - Admin endpoints: 50 requests per minute per admin user

...

    ## Response Format

...
    ```json
    {
...
    }
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: |
        Authenticate user with username/password and return access and refresh tokens.
        Supports optional MFA verification and "remember me" functionality.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              basic:
                summary: Basic login
                value:
                  username: user@example.com
                  password: SecurePassword123!
                  remember_me: false
              withMFA:
                summary: Login with MFA
                value:
                  username: user@example.com
                  password: SecurePassword123!
                  mfa_code: '123456'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimit'
...
  /api/v1/users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Get current authenticated user's complete profile information
      operationId: getUserProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags:
        - Users
      summary: Update user profile
      description: |
        Update current user's profile information. Only provided fields will be updated.
        Some fields like username and email may require additional verification.
      operationId: updateUserProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication. Include the token in the Authorization header:
        ```
        Authorization: Bearer <your-access-token>
        ```
...
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            validationError:
              summary: Validation error example
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: Invalid input parameters
                  details:
                    field: email
                    issue: Invalid email format
                timestamp: '2024-01-20T15:30:45Z'
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            missingToken:
              summary: Missing authentication token
              value:
                success: false
                error:
                  code: MISSING_TOKEN
                  message: Authentication token is required
                  details: null
                timestamp: '2024-01-20T15:30:45Z'
            invalidToken:
              summary: Invalid authentication token
              value:
                success: false
                error:
                  code: INVALID_TOKEN
                  message: The provided authentication token is invalid or expired
                  details: null
                timestamp: '2024-01-20T15:30:45Z'
...
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            rateLimitExceeded:
              summary: Rate limit exceeded
              value:
                success: false
                error:
                  code: RATE_LIMIT_EXCEEDED
                  message: Too many requests. Please try again later.
                  details:
                    retry_after: 60
                timestamp: '2024-01-20T15:30:45Z'
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per time window
          example: 100
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
          example: 0
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when rate limit resets
          example: 1642690305
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
          example: 60
...
  schemas:
    ApiResponse:
      type: object
      description: Standard success response wrapper
      properties:
        success:
          type: boolean
          description: Indicates whether the request was successful
          example: true
        data:
          type: object
          description: Response payload (structure varies by endpoint)
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the response
          example: '2024-01-20T15:30:45Z'
      required:
        - success
        - timestamp
    ApiError:
      type: object
      description: Standard error response wrapper
      properties:
        success:
          type: boolean
          description: Always false for error responses
          example: false
        error:
          type: object
          description: Error details
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: INVALID_CREDENTIALS
            message:
              type: string
              description: Human-readable error message
              example: The provided credentials are invalid
            details:
              type: object
              nullable: true
              description: Additional error context (structure varies by error type)
          required:
            - code
            - message
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the error
          example: '2024-01-20T15:30:45Z'
      required:
        - success
        - error
        - timestamp
...
    LoginRequest:
      type: object
      description: User login credentials
      properties:
        username:
          type: string
          description: Username or email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User password
          example: SecurePassword123!
        remember_me:
          type: boolean
          description: Whether to extend session duration
          default: false
        mfa_code:
          type: string
          description: Multi-factor authentication code (if MFA is enabled)
          example: '123456'
          pattern: ^[0-9]{6}$
      required:
        - username
        - password
    LoginResponse:
      type: object
      description: Successful login response with tokens
      properties:
        access_token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: JWT refresh token for obtaining new access tokens
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          description: Token type (always "Bearer")
          example: Bearer
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 3600
          minimum: 1
        user:
          $ref: '#/components/schemas/UserInfo'
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
        - user
...
    UserInfo:
      type: object
      description: Basic user information included in token responses
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: 123e4567-e89b-12d3-a456-426614174000
        username:
          type: string
          description: Username
          example: user@example.com
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        roles:
          type: array
          description: User roles
          items:
            type: string
          example:
            - user
            - developer
        permissions:
          type: array
          description: User permissions
          items:
            type: string
          example:
            - read:profile
            - write:profile
        mfa_enabled:
          type: boolean
          description: Whether MFA is enabled for this user
          example: false
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: '2024-01-20T15:30:45Z'
        last_login:
          type: string
          format: date-time
          nullable: true
          description: Last login timestamp
          example: '2024-01-20T15:30:45Z'
      required:
        - id
        - username
        - email
        - roles
        - mfa_enabled
        - created_at
...
    UserProfile:
      type: object
      description: Complete user profile information
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: 123e4567-e89b-12d3-a456-426614174000
        username:
          type: string
          description: Username
          example: user@example.com
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        first_name:
          type: string
          nullable: true
          description: User's first name
          example: John
        last_name:
          type: string
          nullable: true
          description: User's last name
          example: Doe
        phone:
          type: string
          nullable: true
          description: Phone number
          example: '+1234567890'
        timezone:
          type: string
          nullable: true
          description: User's timezone
          example: America/New_York
        locale:
          type: string
          nullable: true
          description: User's locale preference
          example: en-US
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: URL to user's avatar image
          example: https://example.com/avatars/user123.jpg
        roles:
          type: array
          description: User roles
          items:
            type: string
          example:
            - user
            - developer
        mfa_enabled:
          type: boolean
          description: Whether MFA is enabled
          example: false
        email_verified:
          type: boolean
          description: Whether email has been verified
          example: true
        account_status:
          type: string
          enum:
            - active
            - suspended
            - locked
            - deleted
          description: Current account status
          example: active
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: '2024-01-20T15:30:45Z'
        updated_at:
          type: string
          format: date-time
          description: Last profile update timestamp
          example: '2024-01-20T15:30:45Z'
        last_login:
          type: string
          format: date-time
          nullable: true
          description: Last login timestamp
          example: '2024-01-20T15:30:45Z'
      required:
        - id
        - username
        - email
        - mfa_enabled
        - email_verified
        - account_status
        - created_at
        - updated_at
    UpdateProfileRequest:
      type: object
      description: User profile update request
      properties:
        first_name:
          type: string
          nullable: true
          description: User's first name
          example: John
        last_name:
          type: string
          nullable: true
          description: User's last name
          example: Doe
        phone:
          type: string
          nullable: true
          description: Phone number
          example: '+1234567890'
        timezone:
          type: string
          nullable: true
          description: User's timezone
          example: America/New_York
        locale:
          type: string
          nullable: true
          description: User's locale preference
          example: en-US
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: URL to user's avatar image
          example: https://example.com/avatars/user123.jpg
...